Fix Direct Descendants Involving Root
This commit fixes a bug in the relative function that caused direct descendants involving a root path (either '', or '/' or '/c:/' or 'c:\\') in either the from or to input to have an erroneous output. 

This bug was fixed in the to -> from direction by adding a direct descendant shortcut check after the inputs are normalized. the check makes sure that from starts with the to input and the next character is a '/' indicating that a new segment is next. 

This bug was fixed in the from -> to direction by making sure that any empty inputs after splitting the segments using '/' are removed.

Unit tests were updated to check for these cases in the appropriate places.