Add fs utility export
This commit adds an export for the fs module to address stubbing issues of node builtin modules. The FileParser class now uses the local fs export to allow stubbing to take place on the utility file instead of the builtin module. This will not affect usage since the fs utility just reexports the node builtin. 

Further, error catching is added to the parse method, so no unhandled rejection warning are emitted when an error occurs.