Add Assertion Plugins
This commit adds a couple of plugins for chai to add more assertions for testing. 

The chai-as-promised package is added to allow easy assertions for async functions. The sinon-chai package is added to help with assertions when using mocks, spies, and stubs through sinon.

Types for these packages are added as well.