Move StreamParser to BrowserStreamParser
This commit renames existing StreamParser implementation to BrowserStreamParser since it uses the Browser Streams API. A version of the Parser will be implemented using the Node Streams API, so it makes sense to separate these two implementations in their names as well.

If any shared functionality exists, an abstract StreamParser will be created that extends Parser and implements the shared functionality.