Add NodeStreamParser
This commit adds an implementation of StreamParser using node streams.

A utility Transform Stream is defined to split the incoming stream of data to a stream of lines. This allows the parse method to be called once every time data appears in the transformed stream instead of having to keep track of splitting lines while parsing.