Split Webpack Entry Points per Environment
This commit adds two new top-level entry points -- one per environment. The webpack config is updated to use those entry points when packing for each environment. 

The top level index exports a combination of both. This is not used for now, but the webpack config will be adjusted when more research is put into the best config to use.