Start Linting Test Files
This commit updates the lint script to start linting test files in addition to the source files. This involved adding an extra tsconfig that includes the test files, then using that config when running the lint commands. 

The lints in the test files were fixed after updating the lint script.