Add NodeStreamParser Unit Tests
This commit adds unit tests for the NodeStreamParser. 

The implementation of NodeStreamParser was updated to add better error handling.

Finally, incoming lines to SummaryBuilder and DetailedSummaryBuilder call `trim()` on the incoming line to make sure the split parts don't include whitespace. This prevents a bug where data isn't tracked properly.