Use webpack for web-only bundle
This commit updates the webpack config to only specify the web environment when bundling. This will be used for script imports, but the other build using tsc will be used to allow proper tree-shaking. 

The entry points are updated to generate a minified version, as well as an unminified version for debug purposes. 

Source maps are changed from inline to a separate file.

The output path is moved to a subdirectory 'web' instead of the base dist folder to prevent overwriting files.

The library name is adjusted to 'VyaptiParser' instead of 'vyapti__parser', mostly for personal preference.