Adjust build scripts
This commit updates the package.json to add better build scripts. The npm-run-all package is used to allow cross-platform script execution.

A separate es2018 version of the package is built using cli options, this is built to the dist/esm2015 to prevent overwriting files. 

Finally, the main js file is updated to use the default esm5 index.js file since that is the default, non-bundled entry-point. Further, other definitions are added for each distribution type.