Rename Dev TSConfig
This commit changes the tsconfig file used for linting to tsconfig.dev.json. This config will be used for development in general since it includes the test files, so it made sense to rename it to something more general.

The base tsconfig now specifically excludes the node_modules folder. 

Finally, the package scripts have been updated to use the new dev config.