Split Test Files based on Environment
This commit updates the test file structure to split files based on the environment the tested files are intended to run in. There will be 3 top level folders in the test/ directory: common/ (for common test suites), node/ (for node only test suites), and browser/ (for browser only test suites). This will allow us to run mocha for the node-only test suites and run karma for the browser-only test suites. Both runners will also run the common test suites.

Coverage reports will need to be generated and combined for both runners.