Update Karma Configuration
This commit updates the karma configuration to remove reporting via karma-typescript (it will get handled by the karma-coverage fork now). The reporter options are removed from the karma typescript config and reporter options are added to the karma-coverage config object. The in-memory option is added so custom merging can be done for the multiple browsers. 

Formatting is also run on the config file.