Add custom coverage scripts
This commit adds custom coverage scripts to manage coverage reports for the project:
- runKarma - a programmatic invocation of karma, so the in-memory reports can be captured for each browser, merged together, then written to one report. 
- combineCoverage - a script that merges the browser and node reports and writes a final combined set of reports.
- mergeCoverage - a helper function that does a more in-depth merge of coverage objects. This is a workaround to instrumentation differences between karma-coverage and nyc. Until those differences are figured out, this allows the two coverage reports to be merged without losing data. 

NOTE: when merging the reports, some erroneous data exists because statements are tracked in the browser instrumentation in node-only files and they aren't tracked in the node instrumentation. Even though they are covered by the node tests, they aren't tracked properly. These small errors are few and far between and the mergeCoverage script workaround is good enough for now.