Add Examples
This commit adds example usages based on the environment. Each parser exposed has an example that includes the async and sync (if applicable) versions of the parse method.

A shared utility function for displaying reports is added so all examples have a similar output.

The examples currently depend on the `dist` file being built, so running the build command is required for examples. This dependency can be removed after the package is published.