Adjust Karma config based on CI env var
This commit updates the karma config to change the browsers list based
on an environment variable. If the CI env var is present, the Karma
config will only add a non-gui browser (ChromeHeadless). This change is
necessary since no display is setup in CI and gui browsers will fail to
launch, causing karma to fail.